Emphasize democratic finance
