+++ src/main/java/com/financedkx/controller/ApiController.java
@@ -1,49 +1,21 @@
-package com.financedkx;
+package com.financedkx.controller;
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.*;
-
-@SpringBootApplication
-public class FinanceDkxApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(FinanceDkxApplication.class, args);
-    }
-}
-
-@Controller
-public class WebController {
-    
-    @GetMapping("/")
-    public String home(Model model) {
-        model.addAttribute("totalSupply", "1,000,000,000");
-        model.addAttribute("currentPrice", "$0.0125");
-        model.addAttribute("marketCap", "$12,500,000");
-        return "index";
-    }
-    
-    @GetMapping("/whitepaper")
-    public String whitepaper() {
-        return "whitepaper";
-    }
-    
-    @GetMapping("/roadmap")
-    public String roadmap() {
-        return "roadmap";
-    }
-}
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/api")
 @CrossOrigin(origins = "*")
 public class ApiController {
-    
+
     @GetMapping("/token-stats")
     public ResponseEntity<Map<String, Object>> getTokenStats() {
         Map<String, Object> stats = new HashMap<>();
@@ -55,11 +27,11 @@ public class ApiController {
         stats.put("circulatingSupply", "750000000");
         return ResponseEntity.ok(stats);
     }
-    
+
     @GetMapping("/transactions")
     public ResponseEntity<List<Map<String, Object>>> getTransactions() {
         List<Map<String, Object>> transactions = new ArrayList<>();
-        
+
         Map<String, Object> tx1 = new HashMap<>();
         tx1.put("hash", "0x1a2b3c4d5e6f...");
         tx1.put("from", "0x742d35cc6434c532532a15c9c0c1b57d");
@@ -67,7 +39,7 @@ public class ApiController {
         tx1.put("amount", "1000.50");
         tx1.put("timestamp", System.currentTimeMillis() - 300000);
         transactions.add(tx1);
-        
+
         Map<String, Object> tx2 = new HashMap<>();
         tx2.put("hash", "0x2b3c4d5e6f7a...");
         tx2.put("from", "0x8ba1f109551bd432803012645hac136c");
@@ -75,7 +47,32 @@ public class ApiController {
         tx2.put("amount", "2500.75");
         tx2.put("timestamp", System.currentTimeMillis() - 600000);
         transactions.add(tx2);
-        
+
         return ResponseEntity.ok(transactions);
+    }
+
+    @GetMapping("/institutions")
+    public ResponseEntity<List<Map<String, Object>>> getInstitutions() {
+        List<Map<String, Object>> institutions = new ArrayList<>();
+
+        institutions.add(Map.of(
+            "name", "Fidelity Investments",
+            "type", "Brokerage",
+            "rating", 4.8
+        ));
+
+        institutions.add(Map.of(
+            "name", "Chase Bank",
+            "type", "Banking",
+            "rating", 4.3
+        ));
+
+        institutions.add(Map.of(
+            "name", "GEICO",
+            "type", "Insurance",
+            "rating", 4.2
+        ));
+
+        return ResponseEntity.ok(institutions);
     }
 }
