+++ src/main/resources/templates/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>FinanceDKX - Decentralizing Finance with DKX Token</title>
+    <title>FinanceDKX - Democratizing Finance for Everyone</title>
     <link rel="stylesheet" href="/css/styles.css">
     <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700&display=swap" rel="stylesheet">
@@ -19,6 +19,7 @@
             <ul class="nav-menu">
                 <li><a href="#home">Home</a></li>
                 <li><a href="#about">About</a></li>
+                <li><a href="/platforms">Platforms</a></li>
                 <li><a href="#tokenomics">Tokenomics</a></li>
                 <li><a href="#roadmap">Roadmap</a></li>
                 <li><a href="#team">Team</a></li>
@@ -37,8 +38,8 @@
     <!-- Hero Section -->
     <section id="home" class="hero">
         <div class="hero-content">
-            <h1 class="hero-title">Decentralize Finance with <span class="gradient-text">DKX</span></h1>
-            <p class="hero-subtitle">The future of decentralized finance is here. Join the revolution that's putting financial power back in your hands.</p>
+            <h1 class="hero-title">Democratize Finance with <span class="gradient-text">DKX</span></h1>
+            <p class="hero-subtitle">The future of inclusive finance is here. Join the movement to put every institution and asset at your fingertips.</p>
             <div class="hero-stats">
                 <div class="stat">
                     <span class="stat-value" th:text="${currentPrice}">$0.0125</span>
@@ -85,8 +86,8 @@
             <h2 class="section-title">About DKX</h2>
             <div class="about-grid">
                 <div class="about-text">
-                    <h3>Revolutionizing Decentralized Finance</h3>
-                    <p>DKX is more than just a cryptocurrency – it's a complete ecosystem designed to democratize finance. Our platform eliminates intermediaries, reduces costs, and gives users complete control over their financial future.</p>
+                    <h3>Revolutionizing Democratic Finance</h3>
+                    <p>DKX is more than cryptocurrency – it's a hub that connects traditional banking, investments, and digital assets. Our goal is to empower everyone with the tools to manage all their finances in one place.</p>
                     <div class="features">
                         <div class="feature">
                             <i class="fas fa-shield-alt"></i>
@@ -196,7 +197,7 @@
             <div class="footer-content">
                 <div class="footer-section">
                     <h4>FinanceDKX</h4>
-                    <p>Decentralizing finance, one transaction at a time.</p>
+                    <p>Democratizing finance, one step at a time.</p>
                     <div class="social-links">
                         <a href="#"><i class="fab fa-twitter"></i></a>
                         <a href="#"><i class="fab fa-telegram"></i></a>
