+++ src/main/resources/templates/platforms.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Financial Platforms Leaderboard - FinanceDKX</title>
+    <title>Democratic Finance Hub - FinanceDKX</title>
     <link rel="stylesheet" href="/css/styles.css">
     <link rel="stylesheet" href="/css/platforms.css">
     <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css" rel="stylesheet">
@@ -32,11 +32,11 @@
         <div class="container">
             <div class="hero-content">
                 <h1 class="hero-title">
-                    Best <span class="gradient-text">Financial Platforms</span> Ranked
+                    Democratizing Finance: <span class="gradient-text">Top Institutions</span>
                 </h1>
                 <p class="hero-subtitle">
-                    Compare banks, investment platforms, credit cards, and financial services. 
-                    Get AI-powered recommendations tailored to your financial goals.
+                    Explore banks, brokers, credit cards and more all in one place.
+                    Get AI-powered recommendations tailored to your goals.
                 </p>
                 <div class="hero-stats">
                     <div class="stat">
